feat(backend): add global and auth roles to attachment routes

+++ server/src/router/attachmentRoutes.ts
@@ -6,7 +6,15 @@ import {
   RequestHandler,
 } from 'express';
 import { CustomRequest } from '../types/CustomRequest';
+import { GlobalRole, ProjectRole } from '@prisma/client';
+import { authorizeGlobalRole } from '../middleware/authorizeGlobalRole';
 import prisma from '../lib/prisma';
+import { checkProjectMembership } from '../middleware/checkProjectMembership';
+import { checkProjectRole } from '../middleware/checkProjectRole';
+import {
+  checkAttachmentOwnership,
+  checkMultipleAttachmentOwnerShip,
+} from '../middleware/checkAttachmentOwnershipMiddleware';
 
 import {
   uploadSingleMiddleware,
@@ -32,6 +40,7 @@ const router = Router();
 // Get all attachments if Entity/Id is not provided
 router.get(
   '/attachments',
+  authorizeGlobalRole(GlobalRole.ADMIN),
   async (req: Request, res: Response): Promise<void> => {
     await getAllAttachments(req, res, prisma);
   }
@@ -40,6 +49,10 @@ router.get(
 // Create single attachment
 router.post(
   '/attachments/single',
+  authorizeGlobalRole(GlobalRole.USER),
+  (req: Request, res: Response, next: NextFunction) =>
+    checkProjectMembership(req as CustomRequest, res, next),
+  checkProjectRole(ProjectRole.USER),
   uploadSingleMiddleware,
   async (req: Request, res: Response, next: NextFunction): Promise<void> => {
     await handleSingleUpload(req as CustomRequest, res, next, prisma);
@@ -49,6 +62,10 @@ router.post(
 // Create several attachments
 router.post(
   '/attachments/many',
+  authorizeGlobalRole(GlobalRole.USER),
+  (req: Request, res: Response, next: NextFunction) =>
+    checkProjectMembership(req as CustomRequest, res, next),
+  checkProjectRole(ProjectRole.USER),
   uploadMultipleMiddleware,
   async (req: Request, res: Response, next: NextFunction): Promise<void> => {
     await handleMultipleUpload(req as CustomRequest, res, next, prisma);
@@ -58,6 +75,11 @@ router.post(
 // Delete attachment
 router.delete(
   '/attachments/:id',
+  authorizeGlobalRole(GlobalRole.USER),
+  (req: Request, res: Response, next: NextFunction) =>
+    checkProjectMembership(req as CustomRequest, res, next),
+  checkProjectRole(ProjectRole.USER),
+  checkAttachmentOwnership(prisma) as RequestHandler,
   validateAttachmentExists as unknown as RequestHandler,
   async (req: Request, res: Response, next: NextFunction): Promise<void> => {
     await deleteAttachment(req as CustomRequest, res, next, prisma);
@@ -67,6 +89,11 @@ router.delete(
 // Delete many attachments
 router.delete(
   '/attachments',
+  authorizeGlobalRole(GlobalRole.USER),
+  (req: Request, res: Response, next: NextFunction) =>
+    checkProjectMembership(req as CustomRequest, res, next),
+  checkProjectRole(ProjectRole.USER),
+  checkMultipleAttachmentOwnerShip(prisma) as RequestHandler,
   deleteManyAttachmentMiddleware,
   async (req: Request, res: Response, next: NextFunction): Promise<void> => {
     await deleteManyAttachments(req, res, next, prisma);
@@ -76,6 +103,9 @@ router.delete(
 // Download attachment
 router.get(
   '/attachments/:id/download',
+  (req: Request, res: Response, next: NextFunction) =>
+    checkProjectMembership(req as CustomRequest, res, next),
+  checkProjectRole(ProjectRole.USER),
   downloadSingleAttachmentMiddleware,
   async (req: Request, res: Response, next: NextFunction): Promise<void> => {
     await downloadSingleAttachment(req, res, next, prisma);
@@ -85,6 +115,9 @@ router.get(
 // Download multiple attachments by Entity/Id
 router.post(
   '/attachments/download',
+  (req: Request, res: Response, next: NextFunction) =>
+    checkProjectMembership(req as CustomRequest, res, next),
+  checkProjectRole(ProjectRole.USER),
   downloadMultipleAttachmentMiddleware,
   async (req: Request, res: Response, next: NextFunction): Promise<void> => {
     await downloadMultipleAttachments(req, res, next, prisma);
